FEAT: Agregando repositorio de Viandas, corrigiendo algunos errores y agregando los imports faltantes

+++ Backend/webienestarAPI/src/main/java/modelos/BecaComedor.java
@@ -2,6 +2,14 @@ package modelos;
 
 import java.time.Year;
 
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
 @Entity
 @Data
 @AllArgsConstructor
@@ -10,7 +18,7 @@ public class BecaComedor{
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     private Boolean ingresos;
     private Enum tipoVivienda;
